SD1-A05: Fix type conversion warning
[git-p4: depot-paths = "//depot/C35/Students/bryantan/SD/": change = 435739]

+++ Engine/Code/Engine/Renderer/Renderer.cpp
@@ -132,7 +132,7 @@ void Renderer::DrawVertexArray(int numVertexes, Vertex const* vertexes) const
 
 void Renderer::DrawVertexArray(std::vector<Vertex> vertexes) const
 {
-	DrawVertexArray(vertexes.size(), vertexes.data());
+	DrawVertexArray(static_cast<int>(vertexes.size()), vertexes.data());
 }
 
 //------------------------------------------------------------------------------------------------
